jparse/Makefile now makes a jparse.a library
This library is used in the main Makefile as well as the jparse Makefile
to link into binaries instead of the various jparse.o and json_*.o
files.

The jparse/Makefile still could use a bit of work in ordering and clean
up but it's a bit better than yesterday.

Add make test rule to jparse/Makefile as well so that running make test
will work from that directory as well (it already worked from
jparse/test_jparse).