Make make silent when nothing to be done
To do this a number of things had to be done as follows:

- Prepend @ to some make -C calls in the main Makefile.

- Add man pages to .PHONY and also the target themselves had to be added
in the Makefiles.

- Remove -v option from cp when copying the man pages to man/.

Now running make a second time will be quiet unless something has
changed that necessitates recompilation or copying man pages over.

There is a change to the dbg/Makefile that has to be put into the dbg
repo which I will do next.