modified as workaround, jstr_test.sh to fix make test failure
As a temporary workaround, jparse/test_jparse/jstr_test.sh was
modified to look for jstrencode and jstrdecode in the top level
directory.  This is NOT the correct place for these tools.
The proper fix will have to wait until Makefile, jparse/Makefile
and jparse/test_jparse/Makefile are improved / modified according
to the new Makefile method.