modified as workaround, jparse/Makefile to fix make release faulure
As a temporary workaround, jparse/Makefile was modified to use use
verge in the .. (top level) directory.  This is NOT the correct
place for verge.  The proper fix will have to wait until Makefile,
jparse/Makefile and jparse/test_jparse/Makefile are improved /
modified according to the new Makefile method.