checkpoint move most of util.c under jparse
As part of the process to make jparse.a, we move nearly all of
util.c (and util.h) under the jparse directory.

Removed the include of limit_ioccc.h by util.c so that
the JSON parser code could become more self contained.

TODO: For those util.o functions that are not used by the JSON
      parser code, we can move such functions elsewhere such
      as under perhaps soup/util_soup.c.

In the case opf dyn_array/dyn_test.c, we put a static function
copy pf parse_verbosity() into that code.

TODO: As a temporary solution, we moved parse_verbosity()
      from util.c to entry_util.c for now.  Functions such
      as this should be moved into a new soup/util_soup.c file.

TODO: From jparse.a and use it elsewhere.