Remove call to ioccc_sanity_chks() in txzchk.c
This function is not strictly needed. The rationale was that if
something is wrong with the core then we should not proceed. But
mkiocccentry does call the function and also calls txzchk. Since no
other tool individually uses the function I don't think txzchk needs to
either. It probably never did. This does also speed (maybe not by a lot
and maybe not even noticeable) txzchk_test.sh as the function would be
called every invocation of txzchk and the function calls
check_utf8_posix_map() and check_location_table(). For
check_location_table() it probably isn't as significant but
check_utf8_posix_map() in particular goes through the very large hmap
table and performs actions on it.