Move man pages to subdirectories based on section
This particularly was needed for the man pages not part of other
directories such as jparse, dbg and dyn_array as I already had done that
(though I did update the Makefiles to refer to the path of the man page
in the empty rules - that is the ones that are `@:').

This still does not install the man pages that are symlinks to other man
pages but make install works now for the man pages that have been
renamed.