Fix the make all repeatedly compiling everything
There might still be some issues but at least now running make all more
than once will try recompiling everything. This is on a different branch
because of some problems I had some of which I'm not sure entirely of
the source.

Specifically one to do with permissions. This was very possibly because
of earlier on trying to do make && sudo make install but then the make
install as root actually compiling the object files again - but it might
not be because make clobber all sometimes (not always) resulted in the
problem again.

Sometimes I got an error trying to link dbg but that was because, it
appears, that a Makefile (not in dbg/) did make dbg instead of dbg.o or
dbg.a. I changed it to dbg.a but in retrospect it might need to be
dbg.o: I'm not sure but noting this in case it is a problem. In
particular it's in test_ioccc/Makefile:

     ../dbg/dbg.a: ../dbg/Makefile
    -       @${MAKE} -C ../dbg dbg extern_liba
    +       @${MAKE} -C ../dbg dbg.a extern_liba

Given that it's extern_liba it probably is okay for dbg.a but just in
case I'm noting it (I didn't see the extern_liba until now as I was
focused on the actual problem).

Some objects were added to .PHONY rules and some were removed. Some
rules ('all') had a no-op @: added to them to silence make when nothing
was to be done.