Fix test_ioccc/ioccc_test.sh wrt jparse_test.sh
The test_ioccc/ioccc_test.sh should NOT run jparse_test.sh on the
test_ioccc/test_JSON files because these files are supposed to be valid
JSON but only be invalid according to the chkentry rules. Since an
invalid json file will be a problem also for chkentry the
chkentry_test.sh script will be able to tell if the files are valid or
not: in the case of the good subdirectory if it is invalid json the test
will fail and the same goes for the bad subdirectory as after a parse
error in json the chkentry tool will not continue. The general.json IS
tested both good and bad and this does work as expected:

    jparse/test_jparse/jparse_test.sh: debug[3]: about to run test that must fail: jparse/jparse -v 0 -J 0 -- ./test_ioccc/test_JSON/general.json/bad/auth.missing-line.92.json >> ./jparse_test.log 2>&1
    syntax error in file ./test_ioccc/test_JSON/general.json/bad/auth.missing-line.92.json at line 91 column 0: empty text
    Warning: jparse/jparse: JSON parse tree is NULL
    ERROR[1]: jparse/jparse: invalid JSON

So all is how it should be.