More work on issue #448
The bug_report.sh script has a new option -x: if specified and no
possible problems and no errors detected the log file will not be kept.
The man page has been updated to account for this change.

The main Makefile has a new action bug_report-tx which runs

    ./bug_report -t -x

which prep.sh now runs as well.

The prep.sh has a new option -l logfile but this is currently unused
(pending discussion). For now if specified the log file is removed and
the script tries to create it so that it will be empty on each
invocation. With this update I discovered that prep.sh does not have a
man page but maybe it should have one.

Removed duplicate prep rule from .PHONY in the main Makefile.

Make the assignment of T_FLAG be "-t" when -t specified (just like the
new X_FLAG variable for -x).