Add -l to only write to log file in bug_report.sh
This is not perfect as there are still some things that will be seen.
Currently at least the following will be seen:

- the output of commands

- the output of hostchk.sh

- the output of git commands

- the output of local files like limit_ioccc.sh

These have to be dealt with differently in a future commit.

The reason it is -l without a file is because the script will always
write to a log file.

Updated the man page but did not add that it is not perfect. Instead the
goal is to make it perfect but at a later time.