Bug fix + make bug_report.sh -l completely silent
To do this required writing some new functions and reworking some things
as well but with these changes bug_report.sh -l is now completely
silent.

test_ioccc/prep.sh runs new Makefile rule bug_report-txl which runs
bug_report.sh with -t -x -l which means to not run the make actions
(prep.sh will have already done this), remove the log file if everything
is okay and make the actions silent (respectively). Note though that
prep.sh is still quite verbose. It does have an option for a log file
but it's not yet used and I'm not even clear yet on the way it's
supposed to work.

There were two bugs in bug_report.sh though one did not seem to be
triggered always. The one that was not triggered in every case was that
because the function exec_command() was empty it was a parse error (it
might also be that I had done that at last and then didn't run it
because nothing else was added).

The other bug was that in some cases a function (write_echo - which
probably needs a better name) was called with too many arguments.  Now
however it allows for more than one arg, printing everything passed to
it.

In prep.sh I fixed a typo with a variable name (in some cases LOG_FILE
was referred to as LOGFILE).