Fix bug_report + add support of log file in prep
In bug_report.sh I left off a place where there could have been some
output even with -l specified.

Just like with bug_report.sh there is a write_echo function and also a
exec_command function in prep.sh.

There seems to be an issue with the make all so that sometimes some
files are not compiled in dbg/ and jparse/ and this is breaking
make prep wrt calling bug_report.sh -t -x -l so for now that action is
disabled.

In the make rule I commented out the invocation of prep.sh with the -l
option until more testing can be done (it seems fine but I need to do
other things and the problem with make all not always working has thrown
me off a bit).