reduce make test verbosity due to the Makefiles
The Makefile verbosity of `make test` has been greatly reduced.

Changed `make test` to be silent by default.  To restore
the previous "more test" move, try:

```sh
make test Q= V=@
```
To further reduce the `make test` verbosity, the test rule
no longer depends on all.  In some cases the test rule tests if
the critical tool is executable.  If it is not, it tries to make it
after which if the tool is still not executable, exits with an
error message.  This is NOT done for all Makefiles, nor for
all things made by make all.

The idea of `make test` is to test the current condition.
The above makes are there to be slightly helpful AND
print nothing if the critical tool already exists.

To proper test the entire tree:

```sh
make clobber all test
```