Enable make verbosity in bug_report.sh
Via the (settable) MAKE_FLAGS variable (by default) the make flags are:

    Q= V=@ MAKE_CD_Q= -w

If you wish to disable these you can do:

    ./bug_report.sh -M ''

Alternatively, if you want to add an option or only disable some of the
actions you can do so like:

    ./bug_report.sh -M 'Q='

for example.

We do NOT recommend using the option when actually reporting a bug but
it is there because there are cases where it's useful to have.

Updated the man page for this option as well as clarifying the -x option
in the EXIT STATUS section.