Make prep.sh (via make prep) quieter
This is done via the -l logfile option I added yesterday. Some output is
still shown but much less is shown.

In prep.sh I changed the text 'Start' to 'START' to match the case of
'PASS' and 'FAIL'.

In the issue https://github.com/ioccc-src/mkiocccentry/issues/448
there's the question of whether it make prep should just have for each
make action show a simpler status like:

    make_action 20 depend OK
    make_action 21 shellcheck ERROR exit code 1
    make_action 22 picky ERROR exit code 2
    make_action 23 all_ref OK

This might be something but with this commit it's already much less
output. Instead you'll see something like:

    =-=-= START: make depend =-=-=

    make -f ./Makefile depend

    =-=-= PASS: make depend =-=-=

Whereas before it was more. Still having a single line might be a good
idea for a future commit.