improve in a few minor ways dyn_array man pages
Added a number of synlinks to the dyn_array.3 man page.

Fixed space nit in dbg/Makefile.

Changed the order of SYNOPSIS and improved some of the
spare blocking of SYNOPSIS items.

Added symlinked man pages to the MAN3_PAGES set of
dbg/Makefile and dyn_alloc/Makefile.

Fixed .gitignore to refer to _/man/_ (instead of just _man_).