Fixes for isolating soup from jparse
Commit 45e18efbb0a729a6ee2e1df9a2e83a76c45951f4 introduced a few
problems that this commit fixes. This commit also makes an enhancement
to help bug_report.sh make a better bug report file.

Because some versions are no longer in the version.h file I updated the
limit_ioccc.sh rule in soup/Makefile to fetch those values. This is not
important for run_flex.sh and run_bison.sh as those files actually have
the versions required. But it's useful for the bug_report.sh script.
It's not needed to have all the versions of the scripts in the
limit_ioccc.sh file because bug_report.sh will run each script to
extract the version. Nevertheless for run_bison.sh and run_flex.sh the
version of each will be in the limit_ioccc.sh file as well as I did not
do a check for MIN_ but rather _VERSION. This could probably be changed
but I did not think of it at the time of doing this and it doesn't seem
that important.

Add to TARGETS limit_ioccc.sh in soup/Makefile. This is because make
clobber removes it and it's needed for some scripts. Actually the main
Makefile had make clobber remove limit_ioccc.sh but the file hasn't been
there but rather under soup/ in some while. Nevertheless in
bug_report.sh I explicitly run make -C soup limit_ioccc.sh to update the
file (to force it being there as we don't run explicitly make -C soup
all - though perhaps make all in the main Makefile now does that
indirectly) prior to retrieving it. I made it an optional check just in
case it fails for some reason. Still it's a pretty useful file to have
and we do warn about it.

The removal of '-l limit.sh' from run_bison.sh and run_bison.sh broke
bug_report.sh because it explicitly used this option.

Important reminder: whenever a path changes one should run bug_report.sh
to make sure nothing breaks.

Rename function run_check_optional in bug_report.sh to
run_optional_check.