Update picky URL + improve URLs in Makefiles
A while back the original picky website was no longer online so I made a
picky GitHub repo which is at https://github.com/xexyl/picky. The
Makefiles now refer to this URL instead of the old one and I also made
the text more consistent with some of the other rules that also refer to
a URL (though it's not entirely finished - there are still some
inconsistencies in text).

I also made _URL variables in the Makefiles for tools that we give a URL
for. This is to make it easier to refer to in the Makefile without
having to update the raw URL in every Makefile. Instead we simply update
the variable and nothing else needs to be changed.