Improve Makefiles (consistency, fixes, other)
Change all lines like

    ${Q} HAVE_FOO="`type -P ${FOO}`"; if [[ -z "$$HAVE_FOO" ]]; then \

to

    ${Q} if ! type -P ${FOO} >/dev/null 2>&1; then \

Change wording of those rules so that they're all the same. This
includes in obtaining the tool, the required to run this rule text and
everything else. If it was fatal that the tool did not exist I kept it
fatal but otherwise it's ignored as it was before.

In the rules that refer to tools change the name of the tool to the
variable that is used to check for the tool itself (except in the URLs).

Due to a yank paste error the missing picky tool referred to seqcexit
rather than picky (not in the URL but in the text).

The rules changed are those for picky, seqcexit, checknr, shellcheck and
depend.

Running make release after these changes works well.