Bug fixes in bug_report.sh
There were three bug fixes but two are connected.

The most important bug fix is that sometimes a command (e.g. printf)
would not ever return a path but then we could not get the version  (or
even try to) because of this. There might have been other commands like
this but the solution is to change command -v to type -P.

The other bugs were in the function get_version_minimal(). I forgot to
check the call of is_exec to is_exec_quiet. If it failed (which would be
unlikely for the one command it's called on - cat) it would be an error
(which might need to be but we already check for its existence
elsewhere). This function as a consequence did not check for if the
command is a file prior to trying to run what(1) or ident(1) on it (this
function does not use strings(1) but the same logic would apply if it
did).