Fix make build_man via make all (${V} specific)
Running make clobber all would not rebuild the man/ subdirectory.
Ironically running make bug_report would. This is because bug_report.sh
enables Makefile debugging. I just opened an issue regarding the
Makefile variable ${V} at
https://github.com/ioccc-src/mkiocccentry/issues/471#issue-1524592580.

The solution in the case of build_man is to change the ${V} to ${Q} but
there might be other places this should be done: not sure but that's why
I opened the issue.

Either way it appears that having @: in a Makefile rule will prevent
anything beyond the : from being read as if it was a comment just like
the shell. Thus any line in the Makefiles that has that variable at the
front will not be executed further.