Fix NULL '%s' directives in util.c
These were uncovered via a warning in fedora. In functions shell_cmd()
and pipe_open() if the exit code indicates that there was an error there
was a problem in that there was allocated memory but at the same time we
don't always abort. I moved the free() to after the dbg() calls but
before returning from the function.

In the case the function is supposed to abort the program the memory is
not freed but the OS should do that.  We can't do that because err() is
supposed to be fatal but this way the argument is not NULL when we print
it.

As for dbg() we now print it properly and then return after freeing the
memory. That being said it might be a good idea to check for the
pointers being NULL first. I can do that in another commit but not now.