Add support for alt_url for authors
After discussion it was thought that a single alternative URL would be
reasonable to allow for authors to have more than one URL. This commit
has a lot of changed files as well as some new files.

I've updated the mkiocccentry_test.sh script to have the correct number
of lines per author. It tests both empty (just hitting enter) URLs and
otherwise. It does not try and check if the user has provided a main URL
before allowing for an alternative URL as this seems like unnecessary
complications to the code.

The test_ioccc/test_JSON/auth.json files were updated and in particular
the good/ subdirectory. What I did not do here is add an alt_url field
to every file under auth.json/bad. This might want to be done but there
are a lot of files to do this with. I did add it to some including one
that is an invalid value but whether all should have at least a null
value alt_url I don't know. Obviously the good/ files have had it added.

Many files under soup/ were changed and it is this that gave me some
trouble. I believe I have it sorted now as make test and make release
both work. Running chkentry on various of the auth.json files also works
as expected.