Add missing alt_url to auth.json/bad files
At least those which did not already have it and those which have a
"url" field. For each "url" there was an "alt_url" added. If it was not
null then alt_url is the same string. If it was null then alt_url is
also null. This made it simpler to do and it serves the purpose.