New bug_report.sh version - get more information
We now collect shell options that are set (shopt -s).

When we're getting the bash version, if the basename is "bash", we get
the values of the variables:

    $BASH_VERSION
    ${BASH_VERSINFO[0-4]} # as in each of the five indices
    $MACHTYPE

A note about the BASH_VERSINFO array. We do not string it up as one
string but rather echo each element one at a time. This is a cosmetic
choice as it would be a long line in the bug report. Additionally for
each one the status is shown by not the element or array but what it
stands for. For instance for ${BASH_VERSINFO[4]} we show something like:

    ## RUNNING: echo ${BASH_VERSINFO[4]}
    release
    ## BASH RELEASE STATUS ABOVE

The $MACHTYPE is the same value as ${BASH_VERSINFO[5]} but to be more
clear we use $MACHTYPE. We still do it if it's the command passed to
get_version is "bash" as it's a bash thing (though maybe other shells
have it - I don't actually know now).

We still try --version after this just in case some or all of the above
information could not be obtained.

An example output of the new information is as follows:

    ## RUNNING: shopt -s
    checkwinsize    on
    cmdhist         on
    complete_fullquote      on
    extquote        on
    force_fignore   on
    globasciiranges on
    globskipdots    on
    hostcomplete    on
    interactive_comments    on
    patsub_replacement      on
    progcomp        on
    promptvars      on
    sourcepath      on
    ## OUTPUT OF shopt -s ABOVE

    ## VERSION CHECK FOR: bash
    ## RUNNING: echo $BASH_VERSION
    5.2.9(1)-release
    ## $BASH_VERSION ABOVE
    ## RUNNING: echo ${BASH_VERSINFO[0]}
    5
    ## BASH MAJOR VERSION ABOVE
    ## RUNNING: echo ${BASH_VERSINFO[1]}
    2
    ## BASH MINOR VERSION ABOVE
    ## RUNNING: echo ${BASH_VERSINFO[2]}
    9
    ## BASH PATCH LEVEL ABOVE
    ## RUNNING: echo ${BASH_VERSINFO[3]}
    1
    ## BASH BUILD VERSION ABOVE
    ## RUNNING: echo ${BASH_VERSINFO[4]}
    release
    ## BASH RELEASE STATUS ABOVE
    ## RUNNING: echo $MACHTYPE
    aarch64-apple-darwin22.1.0
    ## $MACHTYPE ABOVE: BASH SYSTEM TYPE

Of course this is a lot more output but some of this might prove useful
when trying to figure out a problem - even if we should normally get
this information from bash --version (at least as far as the bash
variables go).