Re-enable all disabled warnings as they were fixed
It appears that more recent commits in jparse.[ly] have fixed the
problem of some warnings needing to be disabled (I do recall doing some
of this but I did not test the warnings at the time).

I tested this under macOS, fedora and CentOS so it probably is good. In
CentOS I also don't have the right version for flex or bison so that was
tested indirectly as well.

There are no other warnings disabled in any of the Makefiles so this
task of #87 can probably be ticked as complete.