Move some #includes in dyn_array.c to dyn_array.h
With a simple test it would appear that these #includes are not even
needed but that's because dbg.h also includes them. However dyn_array
will eventually be its own repo so that cannot be relied upon.

As for making sure if there are any unnecessary includes it is stated
that it need not be tested extensively and I believe this is a wise idea
because there are well over 300 #includes in the source code and header
files. That's excluding the shell scripts that have any and other files
that aren't compiled (man pages for instance).

So it's possible that this item can also be ticked under issue #87.