Fix hostchk.sh
With the moving of source files and header files the test compiling of
all the include files this repo uses was broken. Now the hostchk.sh
should check all files except perhaps those in jparse/test_jparse.

Hopefully this is the last functional change / fix needed before other
things can be done but it was a problem I thought of when I was working
on commit b2b83007cbbfabaaabb7311a0c2b116a3c473a78 and it's a pretty
important fix.

Updated hostchk.8 man page to add -Z topdir option.