Improve (consistency) some macro names in util.h
To make it a pattern rather than dependent on the function name I
changed the following macros:

    -#define CALLOC_FAILED_EXIT (-2)                /* invalid exit code - calloc() failure */
    -#define SYSTEM_FAILED_EXIT (-3)                /* invalid exit code - system() failed - returned exit 127 */
    -#define FLUSH_FAILED_EXIT (-4)         /* invalid exit code - fflush() failed */
    -#define NULL_ARGS_EXIT (-5)            /* invalid exit code - function called with a NULL arg */
    +#define EXIT_CALLOC_FAILED (-2)                /* invalid exit code - calloc() failure */
    +#define EXIT_SYSTEM_FAILED (-3)                /* invalid exit code - system() failed - returned exit 127 */
    +#define EXIT_FFLUSH_FAILED (-4)                /* invalid exit code - fflush() failed */
    +#define EXIT_NULL_ARGS (-5)            /* invalid exit code - function called with a NULL arg */

This way one can find the related exit codes in the form of EXIT_
instead of FOO_FAILED_EXIT: one can just look for EXIT_.*_FAILED. Of
course one could have just done git grep FAILED but this does not help
when looking at the files themselves.

A note also is that I changed FLUSH to FFLUSH because the function is
not flush() but fflush().