Remove references to ASAN and valgrind
As discussed on GitHub I have removed all the references to ASAN and
valgrind.

In a future commit there will be a document for static analysis as well
as a document (that's much less thorough - initially if not till the end
of time) on dynamic analysis e.g. with valgrind. The static analysis
document will include compile time warnings etc.