Fix missing prototype (-Wmissing-prototypes)
In the case of find_matching_quote() the function is no longer used so
it was removed from util.c.

In the cases of json_sem_count_chk() and timestr_eq_tstamp() the
functions are used but were missing prototypes. These have been added to
the appropriate files. In json_sem.h:

    extern void json_sem_count_chk(struct json_sem *sem, struct dyn_array *count_err);

and in entry_time.h we now have:

    extern bool timestr_eq_tstamp(char const *timestr, time_t timestamp);