Fix -Wmissing-variable-declarations
In the static_analysis.md I have given several examples because there
are a number of types of problems that had to be resolved where even
adding a fix to one of the cases triggered another warning.

The jnum_gen and jnum_chk have had some changes as well which required
make depend too. Because the variables test_count and test_set in
jnum_test.c are declared in jnum_chk.h I have #included jnum_chk.h in
jnum_test.c. However this meant that we have to have an include guard in
jnum_chk.h as otherwise we would trigger -Wunused-function. It seemed
like unnecessary extra work / overkill to make an entirely new header
file for jnum_test when the variables already existed in jnum_chk.h.
It would not even work very well because they're also used in
jnum_chk.c.