Fix -Wshadow
In the case of jparse/json_parse.c and soup/entry_util.c some variables
were renamed. However as discussed in static_analysis.md there are two
variables we ignore this warning: tbl_name in jsemtblgen.c and txzpath
in txzchk.c. This is because the variable in the file is the same as in
the functions - the static file scope variable is passed directly into
those functions and serve to document the functions as working on the
respective variables.