Address -Wunreachable-code
As noted in static_analysis.md the cases reported that we're aware of
are invalid cases.

In the case of flex / bison generated code we must
not modify the code.

In the case of if statements like

    if ((time_t)-1 > 0) {

the purpose is to check for unsigned time_t.

In the case of foo one it's too easy to break something and in the case
we know of (call to exit()) it does not even matter. I noted foo and
another special file in a section on its own in this commit as well due
to the fact that one of th em (exit()) was triggered by
-Wunreachable-code.

Finally in the case of

    if (BITS_IN_BYTE != 8) {
        err(100, __func__, "BITS_IN_BYTE: %d != 8", BITS_IN_BYTE);
        not_reached();
    }

the purpose of this is to make sure that the BITS_IN_BYTE never is
anything but 8.