Fix 89f8a4 under linux
A funny problem with commit 89f8a4b9d9b6f3533b3577398dbd559f09e27ecc
happens under linux which makes a number of steps in prep.sh fail. This
is because a number of Makefile rules depend on make all. But since a
compilation error occurs under linux due to the change in the commit
they all fail.

The problem is:

    In file included from jnum_test.c:37:0:
    jnum_chk.h:83:13: error: 'quiet' defined but not used [-Werror=unused-variable]
     static bool quiet = false;    /* true ==> quiet mode */
		 ^

This is because jnum_test.c does not use quiet but as described in the
static_analysis.md file to have another header file would be problematic
in another way. The solution therefore is to add another include guard.
Doing this solves the problem.

This raises a question of whether or not the other rules (including some
that do not even work on source files) should depend on make all but
that's another matter entirely.

As part of make prep seqcexit has sequenced the exit codes in util.c.