Make bug_report-tx and bug_report-txl errors fatal
A problem was discovered in that if bug_report.sh finds a problem that
it would not be seen in make prep because the -l option writes to log
file only and because the exit code is ignored in the Makefile all would
appear okay when it is not all okay. Instead you would see a line like:

    make_action 24 bug_report-txl OK

even when there are problems.

The rule bug_report does still ignore errors but it's not used iin a
script. Perhaps it should also not ignore errors but since it does not
write only to a log file it's probably okay.