Fix log file not always created in ioccc_test.sh
The problem was that if one or more files were missing, not the right
file type (regular file or directory) or executable the script would
exit before creating the log file.

Instead we now create the log file first and any of the files
missing/not right file type/not executable are written to the log file.

Additionally we use the exit code 5 for this no matter what the file.
This allows for making it an error of a specific type - one or more
files of the right type missing. There is no man page for this script
(there probably should be) so this is not documented in that way.

If files are missing it would look something like:

    Checking for required files and directories

    ./ioccc_test.sh: ERROR: jparse/test_jparse/jnum_chk file not found
    ./ioccc_test.sh: ERROR: txzchk file not found
    ./ioccc_test.sh: ERROR: chkentry file not found

    ./ioccc_test.sh: ERROR: cannot continue