Address -Wcovered-switch-default
This can be ignored as the purpose of having a default case even
when all enum values are there is it's an error. If someone were to add
to the enum but the switch block was not updated and if the default was
removed it would no longer be an error and would proceed to beyond the
switch which should not happen.