Have some fun throwing tar and feathers :-)
Rename the txz_info struct and variable to tarball. This way one can
look at the code and have things like:

	    ++tarball.total_feathers;

instead of

	    ++txz_info.total_feathers;

which is not nearly as fun - and does not go as well with the joke of
tarring and feathering that I added ages ago.

The function show_txz_info was renamed to show_tarball_info as
show_tarball() would suggest showing the contents - but that's done
indirectly via the tar command that txzchk invokes.

There might be future fun updates here but for now it'll suffice.