Improve JSON semantics tables man pages - #475
In jsemtblgen.8 and jsemcgen.8 all references in the description to the
IOCCC have been removed. As well references to IOCCC specific tools
(e.g. in SEE ALSO) have been removed.

A new symlink jsemcgen.sh.8 pointing to jsemcgen.8 has been added to the
repo and this is also installed via make install. There is an issue that
this will result in checknr (via check_man) rule to run on both files
even though one is only a symlink. Perhaps this should be improved but
it works for now.

Because all_sem_ref.sh is IOCCC specific the references to the IOCCC
have not been removed. In this file we need only an example use.

The issue #475 is not yet complete as these man pages still need at
least one example each (except for jsemtblgen as one should use jsemcgen
instead and this is noted in the man page) but the description is, I
believe, much better. Even though these tools are not specific to the
IOCCC the examples could still be IOCCC specific as it's a working
example.