Address and fix -Wshorten-64-to-32
In most cases this has been fixed (sometimes by a cast and other times
by changing the type) but there are two places where it is TBD if
something will change. As it's well tested for now this warning can be
ignored in these places. The TBD places are in the dyn_array facility
and in the json type structs length (len) variable being signed.
Changing it to unsigned would likely introduce problems elsewhere and it
might not be worth it to fix but this can be decided upon later on.