Address -Wsign-conversion
In some cases this could be fixed by changing the type of a variable. In
other cases it requires a cast. I have not touched dyn_array - this has
to be determined later.

Tested all the recent commits under macOS, fedora and CentOS and all
appears to be okay.

I believe this is the last warning triggered by -Weverything under macOS
but whether other systems will have more warnings I do not yet know.