Further address -Wdouble-promotion
Under fedora linux more code triggered this warning. There were several
kinds of fixes here and we also had to change the jnum_gen tool
slightly.

In some cases we could use a cast. In other cases we could use the L
suffix.

As far as the jnum_gen tool goes we modified it so that when it's a long
double the suffix L is added to the generated code (in jnum_test.c).
This solved this problem:

    jnum_test.c:5474:2: warning: implicit conversion increases floating-point precision: 'double' to 'long double' [-Wdouble-promotion]
	    -214748.3647200000000055,       /* JSON floating point value in long double form */
	    ^~~~~~~~~~~~~~~~~~~~~~~~
    jnum_test.c:5296:2: warning: implicit conversion increases floating-point precision: 'double' to 'long double' [-Wdouble-promotion]
	    -2147483647.099999999977,       /* JSON floating point value in long double form */
	    ^~~~~~~~~~~~~~~~~~~~~~~~

On that note a minor change was made to jparse/jnum_header.c: we had to
change the path of an include:

    -#include "json_parse.h"
    +#include "../json_parse.h"

and of course we used the rule rebuild_jnum_test. There is a decision to
be made on this one though: the main Makefile does not have that rule
and it might be that it should; if so it would possibly change the path
so that we would have to undo this change. This can be determined later.