Address -Wextra-semi-stmt
In the case of jparse.c we ignore this.

In the case of iocccsize.c this is not to be touched by me so I won't.

Whether the latter is to be fixed by Landon is I suppose TBD later. But
it probably doesn't matter anyway and it's arguably a superfluous
warning anyway.