Address -Wdeclaration-after-statement
Since this is before C99 and since we use C11 and since we will be
changing the standard to C17 later this year we do not see this as a
problem. In the code it is:

    uintmax_t forty = no_comment;       /* the value is two too much */

but why we did not initialise no_comment at declaration (therefore
allowing us to not mix them) or else move the variables about we provide
no comment. :-) We won't even provide a comment about whether it is TBD
later if it will be fixed. :-)